refactor: rename the react client error callbacks module
